# docs/swagger.yaml
basePath: /
definitions:
...
  ent.Task:
    properties:
      description:
        description: Description holds the value of the "description" field.
        type: string
      done:
        description: Done holds the value of the "done" field.
        type: boolean
      done_date_time:
        description: DoneDateTime holds the value of the "done_date_time" field.
        type: string
      due_date_time:
        description: DueDateTime holds the value of the "due_date_time" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      notes:
        description: Notes holds the value of the "notes" field.
        type: string
      tags:
        description: Tags holds the value of the "tags" field.
        items:
          type: string
        type: array
    type: object
...
  handlers.AddDirectionSubmissionRequest:
    properties:
      text:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    Task management dashboard API server.
    **Authentication:** Use `POST /users/login`, copy `accessToken` from the response, then click **Authorize** and paste it as the `access_token` cookie value. Alternatively, login from Swagger on the same host and cookies are sent automatically.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Dashboard API
  version: "1.0"
paths:
...
  /add-anonymous-task:
    put:
      consumes:
      - application/json
      description: Creates a simple anonymous task with default values
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Task'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create anonymous task
      tags:
      - tasks
  /aliases/{alias}:
    get:
      consumes:
      - application/json
      description: Returns an alias by its string
      parameters:
      - description: Alias
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AliasModel'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get alias by string
      tags:
      - aliases
  /aliases/container:
    put:
      consumes:
      - application/json
      description: 'Replaces container aliases: removes absent ones and adds new ones'
      parameters:
      - description: Aliases sync request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateContainerAliasesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AliasModel'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sync aliases for container
      tags:
      - aliases
  /aliases/container/{type}/{id}:
    get:
      consumes:
      - application/json
      description: Returns all aliases linked to the given container type and id
      parameters:
      - description: Container type
        in: path
        name: type
        required: true
        type: string
      - description: Container ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AliasModel'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List aliases for container
      tags:
      - aliases
  /aliases/file:
    put:
      consumes:
      - application/json
      description: 'Replaces file aliases: removes absent ones and adds new ones.
        filePath is the logical relative path under FILES_DIR.'
      parameters:
      - description: Aliases sync request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateFileAliasesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AliasModel'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Sync aliases for file
      tags:
      - aliases
...
  /answer-question/{id}:
    post:
      consumes:
      - application/json
      description: Sets the answer for a question
      parameters:
      - description: Question ID
        in: path
        name: id
        required: true
        type: integer
      - description: Answer request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AnswerQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.QuestionFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Answer question
      tags:
      - questions
  /change-tasks-order:
    post:
      consumes:
      - application/json
      description: Changes the order of tasks in a container
      parameters:
      - description: Change tasks order request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangeTasksOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Change tasks order
      tags:
      - tasks
  /container-checks:
    post:
      consumes:
      - application/json
      description: Creates a check entry for the given container
      parameters:
      - description: Check creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddContainerCheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ContainerCheck'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add check to container
      tags:
      - container checks
  /container-checks/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a container check by ID
      parameters:
      - description: Container check ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove container check
      tags:
      - container checks
    patch:
      consumes:
      - application/json
      description: Updates a container check's description
      parameters:
      - description: Container check ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PatchContainerCheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ContainerCheck'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update container check
      tags:
      - container checks
  /container-variables:
    post:
      consumes:
      - application/json
      description: Creates a variable entry for the given container (creates variables
        stack if needed)
      parameters:
      - description: Variable creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddContainerVariableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.ContainerVariables'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add variable to container
      tags:
      - container variables
  /container-variables/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a container variable by ID
      parameters:
      - description: Container variable ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove container variable
      tags:
      - container variables
    patch:
      consumes:
      - application/json
      description: Partially updates a container variable's name and/or value
      parameters:
      - description: Container variable ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PatchContainerVariableRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.ContainerVariables'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update container variable
      tags:
      - container variables
  /directions:
    get:
      consumes:
      - application/json
      description: Returns all directions; use open=true to return only open directions
      parameters:
      - description: Return only open directions
        in: query
        name: open
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Direction'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get directions
      tags:
      - directions
    post:
      consumes:
      - application/json
      description: Creates a new direction with optional parent relationship
      parameters:
      - description: Direction creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.NewDirectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DirectionFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create new direction
      tags:
      - directions
  /directions/{id}:
    get:
      consumes:
      - application/json
      description: Returns a direction by ID with child container IDs
      parameters:
      - description: Direction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DirectionFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get direction by ID
      tags:
      - directions
    patch:
      consumes:
      - application/json
      - application/json
      - application/json
      description: |-
        Partially updates a repetitive task's description and/or notes
        Partially updates a long task's description, notes, and/or done state
        Partially updates a direction's description, notes, and/or closed state
      parameters:
      - description: Repetitive task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PatchContainerByIDRequest'
      - description: Long task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PatchLongTaskByIDRequest'
      - description: Direction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.PatchDirectionByIDRequest'
      produces:
      - application/json
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DirectionFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Patch direction by ID
      tags:
      - repetitive-tasks
      - long-tasks
      - directions
  /directions/{id}/stats:
    get:
      consumes:
      - application/json
      description: Returns aggregated stats entries for a direction including submissions
        and descendant activity
      parameters:
      - description: Direction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DirectionStatsEntry'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get direction stats
      tags:
      - directions
  /directions/{id}/submissions:
    get:
      consumes:
      - application/json
      description: Returns submission records for the given direction (newest first)
      parameters:
      - description: Direction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.DirectionSubmission'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List direction submissions
      tags:
      - directions
    post:
      consumes:
      - application/json
      description: Records a text submission for a direction
      parameters:
      - description: Direction ID
        in: path
        name: id
        required: true
        type: integer
      - description: Submission request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddDirectionSubmissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.DirectionSubmission'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add direction submission
      tags:
      - directions
...
  /files/config:
    get:
      description: Returns whether files are stored encrypted and the configured base
        directory
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/services.FilesConfig'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get files storage config
      tags:
      - files
...
schemes:
- http
- https
security:
- AccessTokenCookie: []
securityDefinitions:
  AccessTokenCookie:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"